merge: develop -> main

See merge request s07-webmobile1-sub2/S07P12A508!60

+++ backend/src/main/java/com/ssafy/backend/api/response/debateBoard/AllBoardRes.java
@@ -10,7 +10,7 @@ import lombok.Setter;
 @Builder
 @ApiModel("AllboardResponse")
 public class AllBoardRes {
-    long board_no;
+    int board_no;
     String board_title;
     int max_applicant;
     boolean board_finished;
